Points Sequences are global. An interface for updating them might be required...
