+++ app/Models/Point.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Models\DirtRally;
+namespace App\Models;
 
 class Point extends \Eloquent
 {
