+++ app/Models/PointsSequence.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Models\DirtRally;
+namespace App\Models;
 
 class PointsSequence extends \Eloquent
 {
