+++ app/Services/Facades/PointSequencesFacade.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace App\Services\Facades;
+
+use App\Services\PointSequences;
+use \Illuminate\Support\Facades\Facade;
+
+class PointSequencesFacade extends Facade {
+    protected static function getFacadeAccessor() { return PointSequences::class; }
+}
